+++ core/src/main/java/org/realityforge/arez/ArezContext.java
@@ -47,7 +47,7 @@ public final class ArezContext
    * Support infrastructure for supporting spy events.
    */
   @Nullable
-  private final SpyImpl _spy = ArezConfig.enableSpy() ? new SpyImpl( this ) : null;
+  private final SpyImpl _spy = Arez.areSpiesEnabled() ? new SpyImpl( this ) : null;
   /**
    * Flag indicating whether the scheduler should run next time it is triggered.
    * This should be active only when there is no uncommitted transaction for context..